ft_putstr_fd

+++ src/ft_putstr_fd.c
@@ -6,7 +6,7 @@
 /*   By: irychkov <irychkov@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/22 12:35:50 by irychkov          #+#    #+#             */
-/*   Updated: 2024/04/22 17:38:56 by irychkov         ###   ########.fr       */
+/*   Updated: 2024/04/29 17:45:47 by irychkov         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	ft_putstr_fd(char *s, int fd)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	if (!s)
